Add send to buffer button (#106)
* Show tracking URL and destination URL for messages

* Add throttling for Click Meter API

* Move after_destroy callback

* Add throttling to deletion of Click Meter links

* Add image support when sending to Buffer

* Fix typos in hash

* Add default time zone for system

* Return posting times as just hours and minutes

* Fix message generation in correct Time Zone

* Save all scheduled date times in PST time zone

* Message generation now in PST

* Implement send to Buffer button

* Add list of dedicated domains from Clickmeter

* Add selected image count

* Remove deletion of Click meter links on non-production env

* Move images tab in front of message templates tab

* Small fix for delayed messages

+++ app/models/click_meter_tracking_link.rb
@@ -19,7 +19,9 @@ class ClickMeterTrackingLink < ActiveRecord::Base
   before_destroy :delete_click_meter_tracking_link
   
   def delete_click_meter_tracking_link
-    ClickMeterClient.delete_tracking_link(click_meter_id)
-    Kernel.sleep(0.1) if Rails.env.production?
+    if Rails.env.production?
+      ClickMeterClient.delete_tracking_link(click_meter_id)
+      Kernel.sleep(0.1) 
+    end
   end
 end
